# SDC/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SDC - Sistema de Defesa Civil API",
        "description": "API RESTful escalável para 100k+ usuários simultâneos.\r\n *                  Suporta webhooks, rate limiting inteligente, processamento assíncrono via Redis,\r\n *                  e múltiplos níveis de priorização de requisições.",
        "contact": {
            "name": "SDC API Support",
            "email": "api@sdc.gov.br"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://sdc.gov.br/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.sdc.gov.br",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/logs/recent": {
            "get": {
                "tags": [
                    "Log Viewer"
                ],
                "summary": "Logs recentes",
                "description": "Retorna os logs mais recentes do sistema",
                "operationId": "logsRecent",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Tipo de log (all, api, webhook, integration, error, performance, security)",
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "enum": [
                                "all",
                                "api",
                                "webhook",
                                "integration",
                                "error",
                                "performance",
                                "security"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de logs a retornar",
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "maximum": 1000
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "logs": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via Bearer Token (Sanctum)",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}